requirements.txt
asgiref==3.4.1
Django==3.2.5
pip==21.1.3
pytz==2021.1
setuptools==57.2.0
sqlparse==0.4.1
beautifulsoup4==4.9.3
behave==1.2.6
behave-django==1.4.0
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
django-db-readonly==0.7.0
django-filter==2.4.0
django-sslserver==0.22
django-webtest==1.9.7
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
flexmock==0.10.4
gprof2dot==2021.2.21
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
memory-profiler==0.58.0
ordereddict==1.1
packaging==21.0
parse==1.19.0
parse-type==0.5.2
psutil==5.8.0
psycopg2==2.9.1
PyJWT==2.1.0
pyparsing==2.4.7
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
uritemplate==3.0.1
urllib3==1.26.6
uWSGI==2.0.19.1
waitress==2.0.0
WebOb==1.8.7
WebTest==2.0.35